FUSETOOLS2-229 - ensure extension is activated before test requiring 'kamel' are executed
it would be nicer if the kamel utilities will handle this case and
auto-activate the exension but I think that in "real" world, outside of
tests, the extension twill always be activbated when it will be accessed
anyway
